fix: useSearchParams 빌드 오류 해결을 위한 컴포넌트 분리
